+++ Docs/Whats_New.md
@@ -23,6 +23,8 @@ Released TBD
 - `vkAllocateDescriptorSets()`: Per Vulkan spec, if any descriptor set allocation 
   fails, populate all descriptor set pointers with `VK_NULL_HANDLE`. In addition, 
   return `VK_ERROR_FRAGMENTED_POOL` if failure was due to pool fragmentation.
+- `vkUpdateDescriptorSets()`: Per Vulkan spec, allow write or copy beyond the 
+  end of a descriptor binding count, including inline uniform block descriptors.
 - Fix rendering issue with render pass that immediately follows a kernel dispatch.
 - Ensure all MoltenVK config info set by `VK_EXT_layer_settings` is used.
 - Move primitive-restart-disabled warning from renderpass to pipeline creation, to reduce voluminous log noise.